sprint 15 v.1.2.

+++ src/main/java/ru/practicum/shareit/booking/BookingServiceImpl.java
@@ -123,7 +123,7 @@ public class BookingServiceImpl implements BookingService {
                 throw new PaginationException("Ошибка пагинации!");
             }
             int page;
-            if (from == size) {
+            if (from.longValue() == size.longValue()) {
                 page = from / size - 1;
             } else {
                 page = from / size;
@@ -186,7 +186,7 @@ public class BookingServiceImpl implements BookingService {
                 throw new PaginationException("Ошибка пагинации!");
             }
             int page;
-            if (from == size) {
+            if (from.longValue() == size.longValue()) {
                 page = from / size - 1;
             } else {
                 page = from / size;
