+++ src/main/java/ru/practicum/shareit/item/ItemServiceImpl.java
@@ -110,7 +110,7 @@ public class ItemServiceImpl implements ItemService {
                 throw new PaginationException("Ошибка пагинации!");
             }
             int page;
-            if (from == size) {
+            if (from.longValue() == size.longValue()) {
                 page = from / size - 1;
             } else {
                 page = from / size;
@@ -142,7 +142,7 @@ public class ItemServiceImpl implements ItemService {
                 throw new PaginationException("Ошибка пагинации!");
             }
             int page;
-            if (from == size) {
+            if (from.longValue() == size.longValue()) {
                 page = from / size - 1;
             } else {
                 page = from / size;
